#42 [MOVE] : StorageViewHolder, Adapter 위치 이동
