+++ app/src/main/java/com/univoice/feature/storage/HomeNoticeAdapter.kt
@@ -1,4 +1,4 @@
-package com.univoice.feature.home
+package com.univoice.feature.storage
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
